Move access granting into package grants
This change moves all grouping and resolving of user grants into a grants
package. This will make it reusable and make the currently large controller
implementation more manageable.
