+++ pkg/controller/postgresqluser/postgresqluser_controller.go
@@ -11,6 +11,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/spf13/pflag"
 	lunarwayv1alpha1 "go.lunarway.com/postgresql-controller/pkg/apis/lunarway/v1alpha1"
+	"go.lunarway.com/postgresql-controller/pkg/grants"
 	"go.lunarway.com/postgresql-controller/pkg/iam"
 	"go.lunarway.com/postgresql-controller/pkg/kube"
 	"go.lunarway.com/postgresql-controller/pkg/postgres"
@@ -70,9 +71,9 @@ func parseFlags(c *ReconcilePostgreSQLUser) {
 	for host := range c.hostCredentials {
 		hostNames = append(hostNames, host)
 	}
-	c.allDatabasesReadEnabled, err = FlagSet.GetBool("all-databases-enabled-read")
+	c.granter.AllDatabasesReadEnabled, err = FlagSet.GetBool("all-databases-enabled-read")
 	parseError(err, "all-databases-enabled-read")
-	c.allDatabasesWriteEnabled, err = FlagSet.GetBool("all-databases-enabled-write")
+	c.granter.AllDatabasesWriteEnabled, err = FlagSet.GetBool("all-databases-enabled-write")
 	parseError(err, "all-databases-enabled-write")
 
 	log.Info("Controller configured",
@@ -82,8 +83,8 @@ func parseFlags(c *ReconcilePostgreSQLUser) {
 		"awsPolicyName", c.awsPolicyName,
 		"awsRegion", c.awsRegion,
 		"awsAccountID", c.awsAccountID,
-		"allDatabasesReadEnabled", c.allDatabasesReadEnabled,
-		"allDatabasesWriteEnabled", c.allDatabasesWriteEnabled,
+		"allDatabasesReadEnabled", c.granter.AllDatabasesReadEnabled,
+		"allDatabasesWriteEnabled", c.granter.AllDatabasesWriteEnabled,
 	)
 }
 
@@ -118,10 +119,16 @@ func Add(mgr manager.Manager) error {
 // newReconciler returns a new reconcile.Reconciler
 func newReconciler(mgr manager.Manager) reconcile.Reconciler {
 	c := &ReconcilePostgreSQLUser{
-		client:           mgr.GetClient(),
-		resourceResolver: kube.ResourceValue,
-		allDatabases:     kube.PostgreSQLDatabases,
-		setAWSPolicy:     iam.SetAWSPolicy,
+		client: mgr.GetClient(),
+		granter: grants.Granter{
+			AllDatabases: func(namespace string) ([]lunarwayv1alpha1.PostgreSQLDatabase, error) {
+				return kube.PostgreSQLDatabases(mgr.GetClient(), namespace)
+			},
+			ResourceResolver: func(resource lunarwayv1alpha1.ResourceVar, namespace string) (string, error) {
+				return kube.ResourceValue(mgr.GetClient(), resource, namespace)
+			},
+		},
+		setAWSPolicy: iam.SetAWSPolicy,
 	}
 	parseFlags(c)
 	return c
@@ -151,21 +158,18 @@ var _ reconcile.Reconciler = &ReconcilePostgreSQLUser{}
 type ReconcilePostgreSQLUser struct {
 	// This client, initialized using mgr.Client() above, is a split client
 	// that reads objects from the cache and writes to the apiserver
-	client           client.Client
-	resourceResolver func(client client.Client, resource lunarwayv1alpha1.ResourceVar, namespace string) (string, error)
-	allDatabases     func(client client.Client, namespace string) ([]lunarwayv1alpha1.PostgreSQLDatabase, error)
-	setAWSPolicy     func(log logr.Logger, credentials *credentials.Credentials, policy iam.AWSPolicy, userID string) error
+	client       client.Client
+	granter      grants.Granter
+	setAWSPolicy func(log logr.Logger, credentials *credentials.Credentials, policy iam.AWSPolicy, userID string) error
 
-	grantRoles               []string
-	rolePrefix               string
-	awsPolicyName            string
-	awsRegion                string
-	awsAccountID             string
-	awsProfile               string
-	awsAccessKeyID           string
-	awsSecretAccessKey       string
-	allDatabasesReadEnabled  bool
-	allDatabasesWriteEnabled bool
+	grantRoles         []string
+	rolePrefix         string
+	awsPolicyName      string
+	awsRegion          string
+	awsAccountID       string
+	awsProfile         string
+	awsAccessKeyID     string
+	awsSecretAccessKey string
 
 	// contains a map of credentials for hosts
 	hostCredentials map[string]postgres.Credentials
@@ -200,7 +204,7 @@ func (r *ReconcilePostgreSQLUser) Reconcile(request reconcile.Request) (reconcil
 	reqLogger = reqLogger.WithValues("user", user.Spec.Name)
 
 	reqLogger.Info("Reconciling PostgreSQLUser", "user", user.Spec.Name)
-	accesses, err := r.groupAccesses(reqLogger, request.Namespace, user.Spec.Read, user.Spec.Write)
+	accesses, err := r.granter.GroupAccesses(reqLogger, request.Namespace, user.Spec.Read, user.Spec.Write)
 	if err != nil {
 		return reconcile.Result{}, fmt.Errorf("group accesses: %w", err)
 	}
@@ -233,7 +237,7 @@ func (r *ReconcilePostgreSQLUser) Reconcile(request reconcile.Request) (reconcil
 	return reconcile.Result{}, nil
 }
 
-func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRoles(log logr.Logger, name string, accesses HostAccess, hosts map[string]*sql.DB) error {
+func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRoles(log logr.Logger, name string, accesses grants.HostAccess, hosts map[string]*sql.DB) error {
 	for host, access := range accesses {
 		connection, ok := hosts[host]
 		if !ok {
@@ -247,7 +251,7 @@ func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRoles(log logr.Logger, name st
 	return nil
 }
 
-func databaseSchemas(accesses []ReadWriteAccess) []postgres.DatabaseSchema {
+func databaseSchemas(accesses []grants.ReadWriteAccess) []postgres.DatabaseSchema {
 	var ds []postgres.DatabaseSchema
 	for _, access := range accesses {
 		ds = append(ds, postgres.DatabaseSchema{
@@ -259,7 +263,7 @@ func databaseSchemas(accesses []ReadWriteAccess) []postgres.DatabaseSchema {
 	return ds
 }
 
-func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRole(log logr.Logger, db *sql.DB, name string, accesses []ReadWriteAccess) error {
+func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRole(log logr.Logger, db *sql.DB, name string, accesses []grants.ReadWriteAccess) error {
 	name = fmt.Sprintf("%s%s", r.rolePrefix, name)
 	log = log.WithValues("operation", "ensurePostgreSQLRole", "name", name)
 	err := postgres.Role(log, db, name, r.grantRoles, databaseSchemas(accesses))
@@ -269,17 +273,7 @@ func (r *ReconcilePostgreSQLUser) ensurePostgreSQLRole(log logr.Logger, db *sql.
 	return nil
 }
 
-// HostAccess represents a map of read and write access requests on host names
-// including the database path.
-type HostAccess map[string][]ReadWriteAccess
-
-type ReadWriteAccess struct {
-	Host     string
-	Database postgres.DatabaseSchema
-	Access   lunarwayv1alpha1.AccessSpec
-}
-
-func (r *ReconcilePostgreSQLUser) connectToHosts(accesses HostAccess) (map[string]*sql.DB, error) {
+func (r *ReconcilePostgreSQLUser) connectToHosts(accesses grants.HostAccess) (map[string]*sql.DB, error) {
 	hosts := make(map[string]*sql.DB)
 	var errs error
 	for hostDatabase := range accesses {
@@ -308,140 +302,4 @@ func (r *ReconcilePostgreSQLUser) connectToHosts(accesses HostAccess) (map[strin
 		hosts[hostDatabase] = db
 	}
 	return hosts, errs
-}
-
-func (r *ReconcilePostgreSQLUser) groupAccesses(reqLogger logr.Logger, namespace string, reads []lunarwayv1alpha1.AccessSpec, writes []lunarwayv1alpha1.AccessSpec) (HostAccess, error) {
-	if len(reads) == 0 {
-		return nil, nil
-	}
-	hosts := make(HostAccess)
-	var errs error
-
-	err := r.groupByHosts(reqLogger, hosts, namespace, reads, postgres.PrivilegeRead, r.allDatabasesReadEnabled)
-	if err != nil {
-		errs = multierr.Append(errs, err)
-	}
-	err = r.groupByHosts(reqLogger, hosts, namespace, writes, postgres.PrivilegeWrite, r.allDatabasesWriteEnabled)
-	if err != nil {
-		errs = multierr.Append(errs, err)
-	}
-
-	if len(hosts) == 0 {
-		return nil, errs
-	}
-	return hosts, errs
-}
-
-func (r *ReconcilePostgreSQLUser) groupByHosts(reqLogger logr.Logger, hosts HostAccess, namespace string, accesses []lunarwayv1alpha1.AccessSpec, privilege postgres.Privilege, allDatabasesEnabled bool) error {
-	var errs error
-	for i, access := range accesses {
-		host, err := r.resourceResolver(r.client, access.Host, namespace)
-		if err != nil {
-			errs = multierr.Append(errs, fmt.Errorf("resolve host: %w", &AccessError{
-				Access: accesses[i],
-				Err:    err,
-			}))
-			continue
-		}
-		if access.AllDatabases {
-			if !allDatabasesEnabled {
-				reqLogger.WithValues("spec", access, "privilege", privilege).Info("Skipping access spec: allDatabases feature not enabled")
-				continue
-			}
-			err := r.groupAllDatabasesByHost(hosts, host, namespace, access, privilege)
-			if err != nil {
-				errs = multierr.Append(errs, fmt.Errorf("all databases: %w", &AccessError{
-					Access: accesses[i],
-					Err:    err,
-				}))
-			}
-			continue
-		}
-		database, err := r.resourceResolver(r.client, access.Database, namespace)
-		if err != nil {
-			errs = multierr.Append(errs, fmt.Errorf("resolve database: %w", &AccessError{
-				Access: accesses[i],
-				Err:    err,
-			}))
-			continue
-		}
-		schema, err := r.resourceResolver(r.client, access.Schema, namespace)
-		if err != nil {
-			errs = multierr.Append(errs, fmt.Errorf("resolve schema: %w", &AccessError{
-				Access: accesses[i],
-				Err:    err,
-			}))
-			continue
-		}
-		hostDatabase := fmt.Sprintf("%s/%s", host, database)
-		hosts[hostDatabase] = append(hosts[hostDatabase], ReadWriteAccess{
-			Host: host,
-			Database: postgres.DatabaseSchema{
-				Name:       database,
-				Schema:     schema,
-				Privileges: privilege,
-			},
-			Access: accesses[i],
-		})
-	}
-	return errs
-}
-
-// groupAllDatabasesByHost groups read write accesses for all known databases in the hosts access map.
-func (r *ReconcilePostgreSQLUser) groupAllDatabasesByHost(hosts HostAccess, host string, namespace string, access lunarwayv1alpha1.AccessSpec, privilege postgres.Privilege) error {
-	databases, err := r.allDatabases(r.client, namespace)
-	if err != nil {
-		return fmt.Errorf("get all databases: %w", err)
-	}
-	var errs error
-	for _, databaseResource := range databases {
-		database := databaseResource.Spec.Name
-		// this limits the `allDatabases` field to only work grant access in a
-		// schema named after the database
-		schema := databaseResource.Spec.Name
-		databaseHost, err := r.resourceResolver(r.client, databaseResource.Spec.Host, namespace)
-		if err != nil {
-			errs = multierr.Append(errs, fmt.Errorf("resolve database '%s' host name: %w", databaseResource.Spec.Name, err))
-			continue
-		}
-		if host != databaseHost {
-			continue
-		}
-		hostKey := fmt.Sprintf("%s/%s", host, database)
-		hosts[hostKey] = append(hosts[hostKey], ReadWriteAccess{
-			Host: host,
-			Database: postgres.DatabaseSchema{
-				Name:       database,
-				Schema:     schema,
-				Privileges: privilege,
-			},
-			Access: access,
-		})
-	}
-	if errs != nil {
-		return errs
-	}
-	return nil
-}
-
-type AccessError struct {
-	Access lunarwayv1alpha1.AccessSpec
-	Err    error
-}
-
-var _ error = &AccessError{}
-
-func (err *AccessError) Error() string {
-	host := err.Access.Host.Value
-	if host == "" && err.Access.Host.ValueFrom.SecretKeyRef != nil {
-		host = fmt.Sprintf("from secret '%s' key '%s'", err.Access.Host.ValueFrom.SecretKeyRef.Name, err.Access.Host.ValueFrom.SecretKeyRef.Key)
-	}
-	if host == "" && err.Access.Host.ValueFrom.ConfigMapKeyRef != nil {
-		host = fmt.Sprintf("from config map '%s' key '%s'", err.Access.Host.ValueFrom.ConfigMapKeyRef.Name, err.Access.Host.ValueFrom.ConfigMapKeyRef.Key)
-	}
-	return fmt.Sprintf("access to host %s: %v", host, err.Err)
-}
-
-func (err *AccessError) Unwrap() error {
-	return err.Err
 }
